+++ src/algorithms/DcdEnergyFromCoulombIntegrals.cxx
@@ -124,8 +124,8 @@ void DcdEnergyFromCoulombIntegrals::iterate(int i) {
       // Slice if Vabcd is not specified
 
       // Read the sliceRank. If not provided use No
-      int64_t sliceRank(getIntegerArgument
-			("sliceRank",No));
+      int sliceRank(getIntegerArgument
+		    ("sliceRank",No));
 
       // Slice loop starts here
       for (int b(0); b < Nv; b += sliceRank) {
@@ -193,8 +193,8 @@ void DcdEnergyFromCoulombIntegrals::dryIterate() {
     DryTensor<> Xakic(4, voov, syms);
 
     // Read the sliceRank. If not provided use No
-    int64_t sliceRank(getIntegerArgument
-		      ("sliceRank",No));
+    int sliceRank(getIntegerArgument
+		  ("sliceRank",No));
 
     if (!Vabcd) {
       // slice if Vabcd is not specified
