+++ src/algorithms/DcsdEnergyFromCoulombIntegrals.cxx
@@ -181,8 +181,8 @@ void DcsdEnergyFromCoulombIntegrals::iterate(int i) {
       // Slice if Vabcd is not specified
 
       // Read the sliceRank. If not provided use No
-      int64_t sliceRank(getIntegerArgument
-			("sliceRank",No));
+      int sliceRank(getIntegerArgument
+		    ("sliceRank",No));
 
       // Slice loop starts here
       for (int b(0); b < Nv; b += sliceRank) {
